+++ text/appetizers/4-events.md
@@ -1,74 +1,107 @@
 # Using Events
 *[`pallets/simple-event`](https://github.com/substrate-developer-hub/recipes/tree/master/pallets/simple-event)*, *[`pallets/generic-event`](https://github.com/substrate-developer-hub/recipes/tree/master/pallets/generic-event)*
 
-In Substrate, [transaction](https://docs.substrate.dev/docs/glossary#section-transaction) finality does not guarantee the execution of functions dependent on the given transaction. To verify that functions have executed successfully, emit an [event](https://docs.substrate.dev/docs/glossary#section-events) at the bottom of the function body.
+Having a [transaction](https://docs.substrate.dev/docs/glossary#section-transaction) included in a block does not guarantee that the function executed successfully. As we saw in the previous recipe, many calls can cause errors, but still be included in a block. To verify that functions have executed successfully, emit an [event](https://docs.substrate.dev/docs/glossary#section-events) at the bottom of the function body.
 
-> **Events** notify the off-chain world of successful state transitions
+Events notify the off-chain world of successful state transitions.
 
-To declare an event, use the [`decl_event`](https://substrate.dev/rustdocs/master/frame_support/macro.decl_event.html) macro.
+## Some Prerequisites
 
-## Simple Event
+When using events, we have to include the `Event` type in our configuration trait. Although the syntax is a bit complex, it is the same every time. If you are a skilled Rust programmer you will recognize this as a series of [trait bounds](https://doc.rust-lang.org/book/ch10-02-traits.html). If you don't recognize this feature of Rust yet, don't worry; it is the same every time, so you can just copy it and move on.
+
+```rust, ignore
+pub trait Trait: system::Trait {
+	type Event: From<Event> + Into<<Self as system::Trait>::Event>;
+}
+```
+
+Next we have to add a line inside of the `decl_module!` macro which generates the 	`deposit_event` function we'll use later when emitting our events. Even experienced Rust programmers will not recognize this syntax because it is unique to this macro. Just copy it each time.
+
+```rust, ignore
+decl_module! {
+	pub struct Module<T: Trait> for enum Call where origin: T::Origin {
+
+		// This line is new
+		fn deposit_event() = default;
+
+		// --snip--
+	}
+}
+```
+
+## Declaring Events
+
+To declare an event, use the [`decl_event!` macro](https://substrate.dev/rustdocs/master/frame_support/macro.decl_event.html). Like any rust enum, Events have names, and can optionally carry data with them. The syntax is slightly different depending on whether the events carry data of primitive types, or generic types from the pallet's configuration trait. These two techniques are demonstrated in the `simple-event` and `generic-event` pallets respectively.
+
+### Simple Events
 
 The simplest example of an event uses the following syntax
 
 ```rust, ignore
 decl_event!(
-    pub enum Event {
-        EmitInput(u32),
-    }
+	pub enum Event {
+		EmitInput(u32),
+	}
 );
 ```
 
-The event is emitted at the bottom of the `do_something` function body:
-
-```rust, ignore
-Self::deposit_event(Event::EmitInput(new_number));
-```
-
-## Events with Generic Types
+### Events with Generic Types
 
 [Sometimes](https://github.com/substrate-developer-hub/recipes/tree/master/pallets/generic-event) events might contain types from the pallet's Configuration Trait. In this case, it is necessary to specify additional syntax
 
 ```rust, ignore
 decl_event!(
-    pub enum Event<T> where AccountId = <T as system::Trait>::AccountId {
-        EmitInput(AccountId, u32),
-    }
+	pub enum Event<T> where AccountId = <T as system::Trait>::AccountId {
+		EmitInput(AccountId, u32),
+	}
 );
 ```
 
-The syntax for `deposit_event` now takes the `RawEvent` type because it is generic over the pallet's configuration trait
+This example also demonstrates how the `where` clause can be used to specify type aliasing for more readable code.
+
+
+## Emitting Events
+
+Events are emitted from dispatchable calls using the `deposit_event` method.
+
+### Simple Events
+
+The event is emitted at the bottom of the `do_something` function body.
+
+```rust, ignore
+Self::deposit_event(Event::EmitInput(new_number));
+```
+
+### Events with Generic Types
+
+The syntax for `deposit_event` now takes the `RawEvent` type because it is generic over the pallet's configuration trait.
 
 ```rust, ignore
 Self::deposit_event(RawEvent::EmitInput(user, new_number));
 ```
 
-*NOTE*: The event described above only wraps `u32` values. If we want/need the `Event` type to contain multiple types from our runtime, then the `decl_event` would use the following syntax
+### Constructing the Runtime
+
+For the first time in the recipes, our pallet has an associated type in its configuration trait. We must specify this type when implementing its trait. In the case of the `Event` type, this is entirely straight forward, and looks the same for both simple events and generic events.
 
 ```rust, ignore
-decl_event!(
-    pub enum Event<T> {
-        ...
-    }
-)
+impl simple_event::Trait for Runtime {
+	type Event = Event;
+}
 ```
 
-In some cases, the `where` clause can be used to specify type aliasing for more readable code
+Events, like dispatchable calls and storage items, requires a slight change to the line in `construct_runtime!`. Notice that the `<T>` is necessary for generic events.
 
 ```rust, ignore
-decl_event!(
-    pub enum Event<T>
-    where
-        Balance = BalanceOf<T>,
-        <T as system::Trait>::AccountId,
-        <T as system::Trait>::BlockNumber,
-        <T as system::Trait>::Hash,
-    {
-        FakeEvent1(AccountId, Hash, BlockNumber),
-        FakeEvent2(AccountId, Balance, BlockNumber),
-    }
-)
+construct_runtime!(
+	pub enum Runtime where
+		Block = Block,
+		NodeBlock = opaque::Block,
+		UncheckedExtrinsic = UncheckedExtrinsic
+	{
+		// --snip--
+		GenericEvent: generic_event::{Module, Call, Event<T>},
+		SimpleEvent: simple_event::{Module, Call, Event},
+	}
+);
 ```
-
-
-*See the next example to use the simple event syntax in the context of verifying successful execution of an [adding machine](./adder.md)*
